refactor: extract gossip protocol to mesh/gossip.rs
Move gossip exchange, peer list management, and announcement collection
out of mesh/mod.rs into mesh/gossip.rs (759 lines).

Includes:
- initiate_gossip, handle_gossip_stream (gossip wire protocol)
- add_peer, remove_peer, update_transitive_peer (peer list mutations)
- collect_announcements (build gossip payload)
- apply_transitive_ann, peer_meaningfully_changed (gossip helpers)
- PeerAnnouncementV0 compat (legacy JSON wire format)
- backfill_legacy_descriptors, merge_demand

mod.rs goes from 4,950 to 4,209 lines. All 808 tests pass.
